+++ test/data/shared_library/inner.cpp
@@ -1,0 +1,9 @@
+#include "inner.hpp"
+
+#include <iostream>
+
+namespace inner {
+  void hello() {
+    std::cout << "hello, library!" << std::endl;
+  }
+}