+++ test/integration/test_library.py
@@ -3,6 +3,8 @@ import os.path
 from . import *
 pjoin = os.path.join
 
+is_mingw = platform_name() == 'windows' and env.builder('c++').flavor == 'cc'
+
 
 class TestLibrary(IntegrationTest):
     def __init__(self, *args, **kwargs):
@@ -31,13 +33,47 @@ class TestLibrary(IntegrationTest):
         self.assertExists(static_library('library'))
 
 
+@unittest.skipIf(is_mingw, 'xfail on mingw (not sure why)')
 class TestSharedLibrary(IntegrationTest):
     def __init__(self, *args, **kwargs):
-        IntegrationTest.__init__(self, 'shared_library', *args, **kwargs)
+        IntegrationTest.__init__(self, 'shared_library', install=True, *args,
+                                 **kwargs)
 
     def test_build(self):
         self.build()
-        self.assertOutput([executable('program')], 'hello, library!\n')
+
+        env = None
+        if platform_name() == 'windows':
+            env = {'PATH': os.pathsep.join([
+                os.path.abspath(self.target_path(output_file(i)))
+                for i in ('outer', 'middle', 'inner')
+            ])}
+        self.assertOutput([executable('program')], 'hello, library!\n',
+                          env=env)
+
+    @skip_if_backend('msbuild')
+    def test_install(self):
+        self.build('install')
+
+        self.assertDirectory(self.installdir, [
+            pjoin(self.bindir, executable('program').path),
+            pjoin(self.libdir, shared_library('inner/inner').path),
+            pjoin(self.libdir, shared_library('middle/middle').path),
+            pjoin(self.libdir, shared_library('outer/outer').path),
+        ])
+
+        os.chdir(self.srcdir)
+        cleandir(self.builddir)
+
+        env = None
+        if platform_name() == 'windows':
+            env = {'PATH': os.pathsep.join([
+                os.path.abspath(os.path.join(
+                    self.libdir, self.target_path(output_file(i))
+                )) for i in ('outer', 'middle', 'inner')
+            ])}
+        self.assertOutput([pjoin(self.bindir, executable('program').path)],
+                          'hello, library!\n', env=env)
 
 
 class TestStaticLibrary(IntegrationTest):