+++ doc/about/changes.md
@@ -23,6 +23,7 @@ in progress
 
 ### Bug fixes
 - Improved logging of syntax errors in `build.bfg` files
+- Fixed usage of nested shared libraries on Linux (via `-rpath-link`)
 
 ---